+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - mongo: `RevokeAccessToken` attempted to delete the access token twice from 
   the datastore leading to `fosite.ErrNotFound` always being returned.
+- mongo: `RevokeRefreshToken` attempted to delete the refresh token twice from 
+  the datastore leading to `fosite.ErrNotFound` always being returned.
 
 ## [v0.18.8] - 2020-06-11
 ### Fixed